Use raises_or_is_caused_by in test_properties
I've moved this useful function into a utilities module.
